Raise ValueError if csv is not df compatible
It is a bad practice to use the `execute_mdx_dataframe` function with `include_attributes` as True, while omitting the PROPERTIES in the MDX query.

It is bad, because the response JSON is arbitrary and depending on what
attributes exist and don't exist in the TM1 model.

Instead of attempting to work with this flawed response JSON,
TM1py must raise an error demanding that the user
requests attributes properly by the use of PROPERTIES in the MDX query.

Solves #507 and #718