refactor after efficiency tests
results show that one of two methods is most performant for ELISANC functionality. The first is through getting the descendants of the ancestor at a derived number of levels below the ancestor. The other is using a recursive tm1drilldown. The first one has overhead in deriving the level of the element and the ancestor but the intersect is generally more efficient, especially if the queried element is not leaf level.

More tests are required to check the validity and performance under varying dimension shapes.