Feature: Allow a list of consolidations to be unwound for update_or_create_hierarchy_from_dataframe()
Currently update_or_create_hierarchy_from_dataframe() allows complete unwinding of hierarchy, but not specific unwinding for consolidations. 

The code change would allow a list of elements to be specified a parameter so selective unwinding is allowed.

This change also replace the param unwind with unwind_all, used for same purpose. Function is updated to be backward compatible so, using unwind would map the value to unwind_all and continue unwinding the entire hierarchy when True.